+++ examples/bench/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @examples/bench
 
+## 0.0.12
+
+### Patch Changes
+
+- Updated dependencies [44958f4]
+  - @palmares/databases@0.1.9
+  - @palmares/server@0.1.7
+  - @palmares/core@0.1.7
+  - @palmares/sequelize-engine@0.1.9
+  - @palmares/express-adapter@0.1.8
+  - @palmares/node-std@0.1.7
+  - @palmares/logging@0.1.7
+  - @palmares/console-logging@0.1.2
+
 ## 0.0.11
 
 ### Patch Changes