+++ examples/server-express-only/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @examples/server-express-only
 
+## 0.0.11
+
+### Patch Changes
+
+- Updated dependencies [44958f4]
+  - @palmares/server@0.1.7
+  - @palmares/core@0.1.7
+  - @palmares/express-adapter@0.1.8
+  - @palmares/node-std@0.1.7
+  - @palmares/logging@0.1.7
+  - @palmares/console-logging@0.1.2
+
 ## 0.0.10
 
 ### Patch Changes