+++ packages/logging/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @palmares/logging
 
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies [44958f4]
+  - @palmares/core@0.1.7
+
 ## 0.1.6
 
 ### Patch Changes